reorganize interface adapters
