+++ src/main/java/interfaceadapters/login/LoginController.java
@@ -1,4 +1,4 @@
-package interfaceadapters.controllers;
+package interfaceadapters.login;
 
 import usecase.login.LoginInputData;
 import usecase.login.LoginInteractor;
