+++ src/main/java/interfaceadapters/news/NewsController.java
@@ -1,7 +1,7 @@
-package interfaceadapters.controllers;
+package interfaceadapters.news;
 
 import usecase.fetch_news.FetchNewsInputBoundary;
-import interfaceadapters.presenters.FetchNewsPresenter;
+import interfaceadapters.news.FetchNewsPresenter;
 import usecase.fetch_news.FetchNewsInputData;
 
 public class NewsController {
