+++ src/main/java/interfaceadapters/trends/TrendsPresenter.java
@@ -1,4 +1,4 @@
-package interfaceadapters.presenters;
+package interfaceadapters.trends;
 
 import usecase.trends.TrendsOutputBoundary;
 import usecase.trends.TrendsOutputData;
